requirements.txt
boto3==1.34.131
Django==4.2.13
django-storages==1.14.3
djangorestframework==3.15.2
psycopg2==2.9.3
Pillow==9.0.0
celery==5.2.7
redis==4.3.4
channels==3.0.3
channels_redis==3.2.0